Refactor keyboard code into new package
